refactor(ios): replace segmented Picker with TabView(.sidebarAdaptable) architecture
