+++ TableProMobile/TableProMobile/Models/ConnectedTab.swift
@@ -1,0 +1,11 @@
+//
+//  ConnectedTab.swift
+//  TableProMobile
+//
+
+internal enum ConnectedTab: String, CaseIterable {
+    case tables
+    case query
+    case history
+    case settings
+}
