+++ TableProMobile/TableProMobile/Views/ConnectionListView.swift
@@ -122,17 +122,15 @@ struct ConnectionListView: View {
                 navigateToPendingConnection(appState.pendingConnectionId)
             }
         } detail: {
-            NavigationStack {
-                if let connection = selectedConnection {
-                    ConnectedView(connection: connection)
-                        .id(connection.id)
-                } else {
-                    ContentUnavailableView(
-                        "Select a Connection",
-                        systemImage: "server.rack",
-                        description: Text("Choose a connection from the sidebar.")
-                    )
-                }
+            if let connection = selectedConnection {
+                ConnectedView(connection: connection)
+                    .id(connection.id)
+            } else {
+                ContentUnavailableView(
+                    "Select a Connection",
+                    systemImage: "server.rack",
+                    description: Text("Choose a connection from the sidebar.")
+                )
             }
         }
         .sheet(isPresented: $showingAddConnection) {
